Move the funnel outcome box into its own file
